requirements.txt
langchain
flask
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
langchain_groq
sentence-transformers==2.4.0